Split styled-components
