+++ src/components/Card/Card.js
@@ -1,11 +1,13 @@
-import { CardContainer, CardCategory, CardName, CardAddress } from './Card.styled.js';
+import { CardContainer } from '../UI/CardContainer.styled';
+import { Paragraph } from '../UI/Paragraph.styled';
+import { HeadlineTwo } from '../UI/HeadlineTwo.styled';
 
 export default function Card({ entries }) {
 	return (
 		<CardContainer>
-			<CardCategory>{entries.category}</CardCategory>
-			<CardName>{entries.name}</CardName>
-			<CardAddress>{entries.address}</CardAddress>
+			<Paragraph>{entries.category}</Paragraph>
+			<HeadlineTwo>{entries.name}</HeadlineTwo>
+			<Paragraph>{entries.address}</Paragraph>
 		</CardContainer>
 	);
 }
