+++ src/components/UI/CardContainer.styled.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const CardContainer = styled.section`
+	background-color: #ffa375;
+	padding: 20px;
+	border-radius: 20px;
+	max-width: 400px;
+	display: flex;
+	flex-direction: column;
+`;
+
+export { CardContainer };
