+++ src/components/UI/HeadlineTwo.styled.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const HeadlineTwo = styled.h2`
+	font-size: 1.7rem;
+	text-transform: uppercase;
+	padding-bottom: 7px;
+`;
+
+export { HeadlineTwo };
