+++ src/components/UI/PageContainer.styled.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const PageContainer = styled.main`
+	margin: 10px 20px 10px 20px;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+`;
+
+export { PageContainer };
