+++ src/components/UI/Paragraph.styled.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const Paragraph = styled.p`
+	font-size: 1.2rem;
+	padding-bottom: 12px;
+`;
+
+export { Paragraph };
